requirements.txt
aiogram
sqlalchemy
aiosqlite
python-dotenv
python-telegram-bot
